requirements.txt
langchain
openai
pypdf
faiss-cpu
chromadb
streamlit
beautifulsoup4
requests
langsmith
langchain-community
langchain-openai
faiss-cpu        # if you’re using FAISS locally
requests
beautifulsoup4
python-dotenv
streamlit

